backend/requirements.txt
aiogram==3.7.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
kerykeion==4.14.0
anthropic==0.28.0
httpx==0.27.0
geopy==2.4.1
timezonefinder==6.5.2
python-multipart==0.0.9
aiohttp==3.9.5
